requirements.txt
absl-py==1.2.0
aiohttp==3.8.3
aiosignal==1.2.0
altair==4.2.0
async-timeout==4.0.2
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
contourpy==1.0.5
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
filelock==3.8.0
fonttools==4.37.4
frozenlist==1.3.1
fsspec==2022.8.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.12.0
google-auth-oauthlib==0.4.6
grpcio==1.49.1
huggingface-hub==0.10.0
hugsvision==0.75.3
idna==3.4
importlib-metadata==5.0.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
multidict==6.0.2
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
protobuf==3.19.6
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.5
pydeck==0.8.0b3
pyDeprecate==0.3.2
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytorch-lightning==1.7.7
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.13.0
tabulate==0.8.10
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
timm==0.6.11
tokenizers==0.12.1
toml==0.10.2
toolz==0.12.0
torch==1.12.1
torchmetrics==0.9.3
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
transformers==4.22.2
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.8.1
